feat: refactor rate limit logging and implement provider responses handler
